+++ apps/Launcher/qml/FileBrowserImage.qml
@@ -1,6 +1,5 @@
 // Copyright (c) 2015-2016, EPFL/Blue Brain Project
 //                          Raphael Dumusc <raphael.dumusc@epfl.ch>
-
 import QtQuick 2.0
 import "style.js" as Style
 
@@ -12,8 +11,14 @@ Item {
         id: placeholder
         anchors.fill: parent
         gradient: Gradient {
-            GradientStop { position: 0.0; color: Style.placeholderTopColor }
-            GradientStop { position: 1.0; color: Style.placeholderBottomColor }
+            GradientStop {
+                position: 0.0
+                color: Style.placeholderTopColor
+            }
+            GradientStop {
+                position: 1.0
+                color: Style.placeholderBottomColor
+            }
         }
         visible: thumbnail.status !== Image.Ready
     }